Add files via upload
I took a course called fundamentals of electrical engineering. All learning was done through the teacher's instructions and our labs. I had a harder time learning the content since we did not have a textbook or other materials to learn from, so I created my own slideshow for the class to teach myself the content.